Move MigMigration metrics to metrics.go
